### 2020-12-11 #2901 @Kevin-Mattheus-Moerman
@huddlej thanks for your submission. Our system is detecting you paper lacks a `Statement of need` section which is a requirement (https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Can you work to add it? Once you've done that you can call the following in a comment here to update the paper:
`@whedon generate pdf`
