### 2020-12-12 #2901 @Maghnuso
> @Maghnuso are you up for another JOSS review?

Hi Mark, yes, I'd be happy to. It's not as if I don't need the practice :-)
I won't be able to touch it for a few days though, if that's ok...
-M
