### 2020-12-12 #2906 @whedon
[REVIEW]: Augur: a bioinformatics toolkit for phylogenetic analyses of human pathogens
**Submitting author:** @huddlej (<a href="http://orcid.org/0000-0002-4250-2063">John Huddleston</a>)
**Repository:** <a href="https://github.com/nextstrain/augur/" target ="_blank">https://github.com/nextstrain/augur/</a>
**Version:** 10.2.0
