### 2020-12-14 e4d520c
Refactor sequences per group functions by type
Splits the original `_calculate_sequences_per_group` function that
supported deterministic and probabilistic sampling into two separate
functions to clarify both the function interfaces and the readability of
the code. Updates tests for the new
`_calculate_fractional_sequences_per_group` function, moves the
Snakemake functional test into a cram test, and adds additional
functional tests for augur filter with subsampling.
