### 2020-12-17 bd004e9
run_shell_command: Improve formatting of command output on error
Instead of unconditionally printing a repr() of the byte-string, decode
to a char-string, strip surrounding whitespace, and if anything is left,
only then print with an indent relative to our own output.

Many commands have their output redirected to a file, so empty output is
not uncommon.  This avoids the terrible:

    ERROR: b''

message and prioritizes the exit status, which we'll always have.
