### 2020-12-17 474847d
run_shell_command: Provide more information about fatal signals
In particular, this offers a suggestion that an out-of-memory condition
may have caused the OS to kill the command.
