+++ tests/util_support/test_shell_command_runner.py
@@ -41,6 +41,23 @@ class TestShellCommandRunner:
                 "Shell exited 5 when running: cmd.*?some error",
             ),
             (
+                subprocess.CalledProcessError(-15, "actual-cmd"),
+                "Shell exited from fatal signal SIGTERM when running: cmd",
+            ),
+            (
+                subprocess.CalledProcessError(128 + 15, "actual-cmd"),
+                "Shell exited from fatal signal SIGTERM when running: cmd",
+            ),
+            (
+                subprocess.CalledProcessError(128 + 9, "actual-cmd"),
+                "Shell exited from fatal signal SIGKILL when running: cmd.*out-of-memory",
+            ),
+            (
+                # Unknown signal doesn't cause exception
+                subprocess.CalledProcessError(-120, "actual-cmd"),
+                "Shell exited -120 when running: cmd",
+            ),
+            (
                 FileNotFoundError(),
                 "Unable to run shell commands using /bin/bash",
             ),
