### 2020-12-18 #2906 @DavidNickle
**Conflict of interest:** I confirm that I have read the JOSS conflict of interest policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
Code of Conduct

**Repository:** The source code for this software is available at the repository url

**License:** The repository containa a plain-text LICENSE file with the contents of an OSI approved software license.

**Contribution and authorship:** The submitting author made major contributions to the software and the full list of paper authors seem appropriate and complete.

**Functionality:** The software is functional and important adding very important knowledge to the public domain

**Installation:** The installation procedure follows the documentation clearly.  I performed all the task in the Zika_Tuotrial

**Functionality:** I confirmed the functional claims of the on the tutorial data as well as some of my own.

**Performance:** No claims here but it is pretty quick!

**A statement of need:** The authors have a clear audience in mine and the paper and the software should be readably available to that audience.

**Installation instructions:** I ran into an error that clearly was my own (may conda app and phython3 where out of date)! Updated and everything worked including all of the dependencies which are all clearly stated.

**Example usage:** I went through the zika_tutorial and all the functionality worked.  I found this to be fantastic exercise and I will be using Nexstrain in my workflows going forward.

**Automated tests:** YES — tests are available
Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. **YES**

**Software paper:** **YES**

Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided? **YES**

**A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? **YES**

**State of the field:** Do the authors describe how this software compares to other commonly-used packages? YES
Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? YES
References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax? **YES**

Lastly, I would like to state that is a piece of scientific coding that is very useful especially to field of phylogenetic and epidemiology. It drives towards the mantra of repeatable results with code that is highly documented and user friendly to the data scientist. The author's have done an exceptional work at sticking to this theme.
