### 2020-12-21 #2906 @majensen
That's bizarre @Maghnuso - I see you are officially in the review so I doubt it that's the problem. Maybe trying closing and opening the browser?
