### 2020-12-21 #2906 @Maghnuso
> That's bizarre @Maghnuso - I see you are officially in the review so I doubt it that's the problem. Maybe trying closing and opening the browser?
If it's helpful @majensen, when I try to accept the invitation again, I get the following message "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account."
