### 2020-12-21 #2906 @Maghnuso
Hi @majensen, all items on my checklist are now completed, so I am happy to accept the paper! Augur was very easy to install, execute and evaluate. The Zika tutorial was fantastic and the paper was well-written and concise. As @DavidNickle stated, this will be a very useful computational tool for epidemiologists going forward that can be more efficiently curated, as the demands of pathogen surveillance increase.
