### 2020-12-22 e97b84e
align: Refer users to the log file when mafft fails
A small but helpful improvement suggested by John Huddleston in #638.
