### 2020-12-22 #2906 @huddlej
Thank you, @majensen, for the great news and thank you @DavidNickle and @Maghnuso for the reviews!

@DavidNickle, do you recall the general error you got during the installation process and how you updated your conda and Python 3 versions? We can use the steps you took to update our documentation as @majensen suggests. For example, did you end up needing to update conda itself like so?

```bash
conda update conda
```

Also, just to doublecheck, did you use [the Local Installation guide for Nextstrain](https://docs.nextstrain.org/en/latest/guides/install/local-installation.html#install-augur-auspice-with-conda) or [the Augur-specific installation guide](https://docs.nextstrain.org/en/latest/guides/install/augur_install.html)?
