+++ setup.py
@@ -52,7 +52,8 @@ setuptools.setup(
         "jsonschema >=3.0.0, ==3.*",
         "packaging >=19.2",
         "pandas >=1.0.0, ==1.*",
-        "phylo-treetime ==0.8.*"
+        "phylo-treetime ==0.8.*",
+        "xopen >=1.0.1, ==1.*"
     ],
     extras_require = {
         'full': [
