+++ tests/test_io.py
@@ -1,0 +1,179 @@
+#!/usr/bin/env python3
+from Bio import SeqIO
+from Bio.Seq import Seq
+from Bio.SeqRecord import SeqRecord
+import bz2
+import gzip
+import lzma
+from pathlib import Path
+import pytest
+import random
+import sys
+
+from augur.io import open_file, read_sequences, write_sequences
+
+
+def random_seq(k):
+    """Generate a single random sequence of nucleotides of length k.
+    """
+    return "".join(random.choices(("A","T","G","C"), k=k))
+
+def generate_sequences(n, k=10):
+    """Generate n random sequences of length k.
+    """
+    return (
+        SeqRecord(Seq(random_seq(k)), id=f"SEQ_{i}")
+        for i in range(1, n + 1)
+    )
+
+@pytest.fixture
+def sequences():
+    return list(generate_sequences(3))
+
+@pytest.fixture
+def sequences_generator():
+    return generate_sequences(3)
+
+@pytest.fixture
+def fasta_filename(tmpdir, sequences):
+    filename = str(tmpdir / "sequences.fasta")
+    SeqIO.write(sequences, filename, "fasta")
+    return filename
+
+@pytest.fixture
+def additional_fasta_filename(tmpdir, sequences):
+    filename = str(tmpdir / "additional_sequences.fasta")
+    SeqIO.write(sequences, filename, "fasta")
+    return filename
+
+@pytest.fixture
+def gzip_fasta_filename(tmpdir, sequences):
+    filename = str(tmpdir / "sequences.fasta.gz")
+
+    with gzip.open(filename, "wt") as oh:
+        SeqIO.write(sequences, oh, "fasta")
+
+    return filename
+
+@pytest.fixture
+def bzip2_fasta_filename(tmpdir, sequences):
+    filename = str(tmpdir / "sequences.fasta.bz2")
+
+    with bz2.open(filename, "wt") as oh:
+        SeqIO.write(sequences, oh, "fasta")
+
+    return filename
+
+@pytest.fixture
+def lzma_fasta_filename(tmpdir, sequences):
+    filename = str(tmpdir / "sequences.fasta.xz")
+
+    with lzma.open(filename, "wt") as oh:
+        SeqIO.write(sequences, oh, "fasta")
+
+    return filename
+
+@pytest.fixture
+def genbank_reference():
+    return "tests/builds/zika/config/zika_outgroup.gb"
+
+
+class TestReadSequences:
+    def test_read_sequences_from_single_file(self, fasta_filename):
+        sequences = read_sequences(fasta_filename, format="fasta")
+        assert len(list(sequences)) == 3
+
+    def test_read_sequences_from_multiple_files(self, fasta_filename, additional_fasta_filename):
+        sequences = read_sequences(fasta_filename, additional_fasta_filename, format="fasta")
+        assert len(list(sequences)) == 6
+
+    def test_read_sequences_from_multiple_files_or_buffers(self, fasta_filename, additional_fasta_filename):
+        with open(fasta_filename) as fasta_handle:
+            sequences = read_sequences(fasta_handle, additional_fasta_filename, format="fasta")
+            assert len(list(sequences)) == 6
+
+    def test_read_single_fasta_record(self, fasta_filename):
+        record = next(read_sequences(fasta_filename, format="fasta"))
+        assert record.id == "SEQ_1"
+
+    def test_read_single_genbank_record(self, genbank_reference):
+        reference = next(read_sequences(genbank_reference, format="genbank"))
+        assert reference.id == "KX369547.1"
+
+    def test_read_single_genbank_record_from_a_path(self, genbank_reference):
+        reference = next(read_sequences(Path(genbank_reference), format="genbank"))
[…]
