+++ pytest.python3.ini
@@ -1,4 +1,9 @@
 [pytest]
+# ignore biopython's deprecation warnings about alphabet that are outside of our control
+filterwarnings =
+    ignore::PendingDeprecationWarning:Bio.Alphabet
+    ignore:Coverage disabled
+
 addopts =
     # do not capture any output---necessary for interactive breakpoints
     -s
