+++ augur/index.py
@@ -24,24 +24,23 @@ def run(args):
 
     #count occurences and write to output file
     with open(args.output, 'wt') as out_file:
-    	tsv_writer = csv.writer(out_file, delimiter = '\t')
-    	tsv_writer.writerow(['sequence_name', 'length', '#A','#C', '#G', '#T' ,'#N', '#non-ACTGN'])
+        tsv_writer = csv.writer(out_file, delimiter = '\t')
+        tsv_writer.writerow(['sequence_name', 'length', '#A','#C', '#G', '#T' ,'#N', '#non-ACTGN'])
 
-    	for record in seqs:
-    		#change sequences to lower case and remove gaps
-    	   	seq = record.seq.ungap(gap='-').lower()
-    	   	l = len(seq)
+        for record in seqs:
+            #change sequences to lower case and remove gaps
+            seq = record.seq.ungap(gap='-').lower()
+            l = len(seq)
 
-    	   	#count bases
-    	   	a = seq.count('a')
-    	   	c = seq.count('c')
-    	   	g = seq.count('g')
-    	   	t = seq.count('t')
-    	   	n = seq.count('n')
-    	   	non_acgtn = l-sum([a,c,g,t,n])
+            #count bases
+            a = seq.count('a')
+            c = seq.count('c')
+            g = seq.count('g')
+            t = seq.count('t')
+            n = seq.count('n')
+            non_acgtn = l-sum([a,c,g,t,n])
 
-    	   	#write row for each sequence
-    	   	tsv_writer.writerow([record.id,l,a,c,g,t,n,non_acgtn])
+            #write row for each sequence
+            tsv_writer.writerow([record.id,l,a,c,g,t,n,non_acgtn])
 
     seqs.close()
-    out_file.close()
