### 2020-12-29 e40c5c8
added functionality to count other IUPAC names and filter out sequences with non nucleotide characters
