### 2020-12-30 dc20ba7
simplify counting of nucleotides
