### 2020-12-30 7bd9106
Fixes parsing of upper case sequences.
