### 2020-12-31 #2906 @DavidNickle
I had to do "conda update conda" and then make sure I had py3 was updated too.  I installed both ways, that is to say Nexstrain and Augur specific install.  I ran the Zika tutorial through the Nextstrain install.  All in all it really was not a big issue.  You guys are doing great work and I think this really helps people with reproducible code and analyses.
