### 2020-12-31 a2ee034
Prototype augur filter sequence index interface
Adds an optional `--sequence-index` argument to augur filter that
accepts the output of the new `augur index` command and uses these
sequence composition data to filter sequences by length or invalid
nucleotide composition instead of inspecting sequences directly.

This prototype interface is backwards-compatible with previous workflows
that don't precalculate the sequence index. As such, it tries to
calculate the index in the fly when one is not provided by the user.
This commit stubs out the functions in `index.py` that would allow this
on-the-
[…]
