+++ augur/filter.py
@@ -6,10 +6,13 @@ from Bio import SeqIO
 from collections import defaultdict
 from typing import Collection
 import random, os, re
+import pandas as pd
 import numpy as np
 import sys
 import datetime
 import treetime.utils
+
+from .index import index_sequences
 from .utils import read_metadata, get_numerical_dates, run_shell_command, shquote, is_date_ambiguous
 
 comment_char = '#'
@@ -89,6 +92,7 @@ def filter_by_query(sequences, metadata_file, query):
 def register_arguments(parser):
     parser.add_argument('--sequences', '-s', required=True, help="sequences in fasta or VCF format")
     parser.add_argument('--metadata', required=True, metavar="FILE", help="sequence metadata, as CSV or TSV")
+    parser.add_argument('--sequence-index', help="sequence composition report generated by augur index. If not provided, an index will be created on the fly.")
     parser.add_argument('--min-date', type=numeric_date, help="minimal cutoff for date; may be specified as an Augur-style numeric date (with the year as the integer part) or YYYY-MM-DD")
     parser.add_argument('--max-date', type=numeric_date, help="maximal cutoff for date; may be specified as an Augur-style numeric date (with the year as the integer part) or YYYY-MM-DD")
     parser.add_argument('--min-length', type=int, help="minimal length of the sequences")
@@ -141,16 +145,28 @@ def run(args):
     #If VCF, open and get sequence names
     if is_vcf:
         seq_keep, all_seq = read_vcf(args.sequences)
-
-    #if Fasta, read in file to get sequence names and sequences
     else:
-        try:
-            seqs = SeqIO.index(args.sequences, 'fasta')
-        except ValueError as error:
-            print("ERROR: Problem reading in {}:".format(args.sequences))
-            print(error)
-            return 1
-        seq_keep = list(seqs.keys())
+        # If FASTA, try to load the sequence composition details and strain
+        # names to be filtered. Generate the sequence index on the fly, for
+        # backwards compatibility with older workflows that don't generate the
+        # index ahead of time.
+        if args.sequence_index is None:
+            # Create a new index based on the given sequence's path.
+            sequence_index_path = args.sequences + ".idx"
+            index_sequences(args.sequences, sequence_index_path)
+        else:
+            sequence_index_path = args.sequence_index
+
+        sequence_index = pd.read_csv(
+            sequence_index_path,
+            sep="\t"
+        )
+
+        # Calculate summary statistics needed for filtering.
+        sequence_index["ACGT"] = sequence_index.loc[:, ["A", "C", "G", "T"]].sum(axis=1)
+        sequence_index["valid_nucleotides"] = sequence_index.loc[:, ["ACGT", "N", "other_IUPAC"]].sum(axis=1)
+        sequence_index["invalid_nucleotides"] = sequence_index["length"] - sequence_index["valid_nucleotides"]
+        seq_keep = sequence_index["sequence_name"].values
         all_seq = seq_keep.copy()
 
     try:
@@ -228,12 +244,13 @@ def run(args):
         if is_vcf: #doesn't make sense for VCF, ignore.
             print("WARNING: Cannot use min_length for VCF files. Ignoring...")
         else:
-            seq_keep_by_length = []
-            for seq_name in seq_keep:
-                sequence = seqs[seq_name].seq
-                length = sum(map(lambda x: sequence.count(x), ["a", "t", "g", "c", "A", "T", "G", "C"]))
-                if length >= args.min_length:
-                    seq_keep_by_length.append(seq_name)
+            is_in_seq_keep = sequence_index["sequence_name"].isin(seq_keep)
[…]
