### 2020-12-31 97cee79
Add instructions to update conda prior to installation
As with the main Nextstrain docs [1], the Augur docs need to mention how
to update conda for users who have an older version installed.

[1] https://github.com/nextstrain/docs.nextstrain.org/pull/41/commits/178a0829179848def908d9a8699cad359528ce8e
