### 2021-01-01 #2906 @huddlej
Thank you for clarifying, @DavidNickle! I've updated both the [Nextstrain installation docs](https://github.com/nextstrain/docs.nextstrain.org/commit/178a0829179848def908d9a8699cad359528ce8e) and [Augur installation docs](https://github.com/nextstrain/augur/pull/655/commits/97cee791938b32ff6a29307b7c0809d995c9e34a) to reflect this issue.
