### 2021-01-01 #2906 @majensen
@huddlej can you have a look at the "Missing DOIs" in the above (https://github.com/openjournals/joss-reviews/issues/2906#issuecomment-753240297) and see if they are appropriate to incorporate into your bib references? Thanks
