### 2021-01-01 #2906 @majensen
@huddlej I would like you to consider adding the following committers (not @'ed so they don't get pinged) explicitly to the author list, and then remove the "Augur Contributors" author and make a general acknowledgement in the paper instead.

- kairstenfay
- elebow
- joverlee521
- willingc
- barneypotter24
and possibly
- lmoncla
- sidneymbell
- saikirankv

My rationale is that these folks (many of whom committed more lines or more times that one of your current authors) appear to have contributed:
1. Functionality for new diseases
2. Significant documentation
3. Updates to the use of python and its ecosystem that make the application less homebrew and more pro.

Please consider giving this credit where it is due. Certainly some of that work was free (or let's say, paid forward!). Up to you and your team - I won't dig in my heels.

thanks for your consideration
