### 2021-01-01 96f61be
Add missing DOIs to paper as recommended by JOSS
As recommended by the JOSS reference check [1].

[1] https://github.com/openjournals/joss-reviews/issues/2906#issuecomment-753240297
