+++ paper/paper.bib
@@ -139,7 +139,9 @@ url={https://doi.org/10.1038/s41591-020-0935-z}
 	Pages = {3546--3548},
 	Title = {{nextflu: real-time tracking of seasonal influenza virus evolution in humans.}},
 	Volume = {31},
-	Year = {2015}}
+	Year = {2015},
+    doi = {10.1093/bioinformatics/btv381}
+}
 
 @article{Hadfield2018,
 author = {Hadfield, J. and Megill, C. and Bell, S. M. and Huddleston, J. and Potter, B. and Callender, C. and Sagulenko, P. and Bedford, T. and Neher, R. A.},
@@ -227,7 +229,8 @@ year = {2012},
 volume = {28},
 number = {19},
 pages = {2520--2522},
-month = aug
+month = aug,
+doi = {10.1093/bioinformatics/bts480}
 }
 
 @article{Sagulenko2018,
