### 2021-01-01 #2906 @huddlej
@majensen The missing DOIs were both correct and helpful. I've added these to the bibliography.

I'll check with the rest of the team about the authorship changes. We weren't sure how to best choose a threshold for authorship and I think folks will be open to these changes. We would also want to consider [danielsoneg](https://github.com/nextstrain/augur/search?q=danielsoneg&type=commits) and groutr in this list, if we go this route.
