### 2021-01-01 17caf1c
Update change log in preparation for feature release
