### 2021-01-04 d856cfa
restructure index.py file to match the proposed structure of having two functions index_sequences and index_sequence.
