+++ augur/index.py
@@ -5,10 +5,11 @@ Count occurrence of bases in a set of sequences
 from Bio import SeqIO
 import Bio.Seq
 import Bio.SeqRecord
+import sys
 import csv
 
 def register_arguments(parser):
-    parser.add_argument('--sequences', '-s', required=True, help="sequences in fasta or VCF format")
+    parser.add_argument('--sequences', '-s', required=True, help="sequences in fasta format")
     parser.add_argument('--output', '-o', help="output file", required=True)
 
 
@@ -48,7 +49,10 @@ def index_sequence(sequence, values):
     ['seq_C', 11, 1, 1, 1, 1, 1, 0, 0, 0, 5]
 
     """
-    pass
+    counts = []
+    for v in values:
+        counts.append(sum(map(lambda x: sequence.count(x), v)))
+    return counts
 
 
 def index_sequences(sequences_path, sequence_index_path):
@@ -66,67 +70,52 @@ def index_sequences(sequences_path, sequence_index_path):
         sequence in the given input file.
 
     """
-    pass
+    #read in files
+    try:
+        seqs = SeqIO.parse(sequences_path, 'fasta')
+    except ValueError as error:
+        print("ERROR: Problem reading in {}:".format(sequences_path), file=sys.stderr)
+        print(error, file=sys.stderr)
+        return 1
+    
+    other_IUPAC = ['r', 'y', 's', 'w', 'k', 'm', 'd', 'h', 'b', 'v']
+    values = [['a'],['c'],['g'],['t'],['n'],other_IUPAC,['-'],['?']] 
+    labels = ['A','C','G','T','N','other_IUPAC','-','?']
+    
+    tot_length = 0
+    num_of_seqs = 0
+    
+    with open(sequence_index_path, 'wt') as out_file:
+        tsv_writer = csv.writer(out_file, delimiter = '\t')
+        
+        #write header
+        header = ['strain', 'length']+labels+['invalid_nucleotides']
+        tsv_writer.writerow(header)
+        
+        for record in seqs:
+
+            #change sequences to lower case
+            sequence = record.seq.lower()
+            l = len(sequence)   
+            
+            #index the sequence and write row
+            counts = index_sequence(sequence, values)
+            invalid_nucleotides = l-sum(counts)
+            row = [record.id, l]+counts+[invalid_nucleotides]
+            tsv_writer.writerow(row)
+            
+            tot_length += l 
+            num_of_seqs += 1 
+    
+    seqs.close()    
+    print("Analysed %i sequences with an average length of %i nucleotides."%(num_of_seqs,int(tot_length/num_of_seqs)))
 
 
 def run(args):
     '''
-    counts the number of a,c,t,g,n and other IUPAC nucleotides in a set of sequences
+    runs index_sequences which counts the number of A, C, T, G, N, other IUPAC nucleotides, ambiguous bases
+    ("?" and "-"), and other invalid characters in a set of sequences and write
+    the composition as a data frame to the given sequence index path.
     '''
-    #read in files
-    try:
-        seqs = SeqIO.parse(args.sequences, 'fasta')
-    except ValueError as error:
-        print("ERROR: Problem reading in {}:".format(args.sequences))
-        print(error)
-        return 1
-
-    good_chars = {'a', 'c', 'g', 't', 'n', '-', 'r', 'y', 's', 'w', 'k', 'm', 'd', 'h', 'b', 'v','?'}
-    tot_length = 0
-    num_of_seqs = 0
-    #characters to count as list of strings
-    other_IUPAC = ['r', 'y', 's', 'w', 'k', 'm', 'd', 'h', 'b', 'v']
-    values = [['a'],['c'],['g'],['t'],['n'],other_IUPAC]
[…]
