### 2021-01-04 91a3c70
changes the output of sequence_index to contain strain and length. Makes augur filter compatible with new index output structure. The sequence_name column is now called strain and index now calculates the number of invalid nucleotides making the calculation during filter redundant.
