### 2021-01-05 afae81b
Acknowledge specific contributors
Instead of only linking to the project's contributor statistics on
GitHub, we thank specific contributors by name. Some of these folks do
not appear in the GitHub statistics for annoying technical reasons and
this approach allows us to acknowledge those people.

This commit also removes the placeholder author for open source
contributors since these folks did not sign up for authorship
responsibilities and are now acknowledged more explicitly.
