### 2021-01-05 3f89ded
Add Nextstrain team members who contributed to Augur
