### 2021-01-05 #2906 @huddlej
@majensen, we've decided that the best way to acknowledge all contributors without adding unwanted responsibilities to people outside of our team was to add Nextstrain team members as authors and name all other contributors specifically in the acknowledgements.

To this end, we added three authors from the Nextstrain team and added 25 contributors to the acknowledgements. Since we were able to explicitly acknowledge everyone who has contributed code to this project, we removed the "Augur Contributors" item from the author list. All of these changes are now reflected in `master` of our augur repository.

Thank you for asking us to reconsider our authorship and acknowledgements approach here. This has changed my perspective a bit on how to evaluate contributions to a project.
