### 2021-01-05 e5200b7
index: Correct expected output in doctests
