### 2021-01-05 ebce022
index: Raise ValueErrors if any required conditions are false
The character-counting code makes several assumptions about the
character sets provided to index_sequence(), which, if violated, would
result in incorrect output.

Validation in index_sequence() instead of index_sequences() so that both
can be used safely as an internal API.  The additional per-sequence
overhead appears to be roughly 0.0025ms, which even at 300k sequences is
less than a second.
