### 2021-01-05 9c52421
Use probabilistic sampling by default
Makes the optional flag `--probabilistic-sampling` enabled by default
and adds a `--no-probabilistic-sampling` flag to disable this default
behavior. These changes make the `--probabilistic-sampling` flag
unnecessary but maintain backwards-compatibility for workflows that
reference this flag. The functional tests for augur filter now reflect
this new behavior.
