### 2021-01-05 e1abcbb
Use probabilistic sampling by default
Makes the optional flag `--probabilistic-sampling` enabled by default
and adds a `--no-probabilistic-sampling` flag to disable this default
behavior. These changes make the `--probabilistic-sampling` flag
unnecessary but maintain backwards-compatibility for workflows that
reference this flag.

We toggle probabilistic sampling through one variable. This simplifies
the downstream logic that checks for whether to use probabilistic
sampling or not. Note that with this approach we cannot define a
separate `help` string for the `--no-probabilistic-sampling` flag
[…]
