### 2021-01-05 9d21aad
Warn users that `--probabilistic-sampling` is deprecated
Adds text to the flag's help and also issues a warning to users who
explicitly request the flag. Note that we remove the default `True`
value from the flag so we know when users have requested this option.
