### 2021-01-05 291fc09
index: Elide unstable output from doctest
Otherwise this test fails stochastically.
