### 2021-01-06 01316b2
Report index results from CLI only
Report the number of sequences and their average length only when
indexing sequences from the CLI. This change allows the internal
function `index_sequences` to be called without printing to stdout.
Instead, this function returns the details necessary to provide the
report to the user from `run`.
