### 2021-01-06 cf8a655
Expand description of index output file
For ourselves and other users, it is handy to know generally what the
output from this command will look like.
