### 2021-01-06 fe85c80
Add warning when sequence index is not provided
