### 2021-01-06 #2906 @majensen
@huddlej Thanks very much for your thoughtful consideration; I'm certainly pleased with the team's decision.

Now, the paper looks great. I have no edits, except for the attached patch, which should fix the Alm reference so that it is not whacked by BibTeX.

Before I perform the incantation to summon the Associate Ed-in-Chief, can I have you create an open archive of your repo on Zenodo, Figshare, or similar (see https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729 for a tutorial if needed). Then report the resulting DOI back here in the thread. JOSS strongly prefers that the authors and the title of the archive match those of the paper.

Thanks!

[bibpatch.dif.txt](https://github.com/openjournals/joss-reviews/files/5773769/bibpatch.dif.txt)
