### 2021-01-06 5a190e1
Fix citation for manuscript
Place "Group" as the last name of the final citation for the consortium
so it appears in the proper order in the final bibliography.
