+++ paper/paper.bib
@@ -40,7 +40,7 @@
     number = {32},
     urldate = {2020-08-26},
     journal = {Eurosurveillance},
-    author = {Alm, Erik and Broberg, Eeva K. and Connor, Thomas and Hodcroft, Emma B. and Komissarov, Andrey B. and Maurer-Stroh, Sebastian and Melidou, Angeliki and Neher, Richard A. and O’Toole, Áine and Pereyaslov, Dmitriy and Group, The WHO European Region sequencing laboratories {and} GISAID EpiCoV},
+    author = {Alm, Erik and Broberg, Eeva K. and Connor, Thomas and Hodcroft, Emma B. and Komissarov, Andrey B. and Maurer-Stroh, Sebastian and Melidou, Angeliki and Neher, Richard A. and O’Toole, Áine and Pereyaslov, Dmitriy and The WHO European Region Sequencing Laboratories {and} GISAID EpiCoV Group,},
     month = aug,
     year = {2020},
     note = {Publisher: European Centre for Disease Prevention and Control},
