### 2021-01-06 e9c0f79
Update change log for patch release
