### 2021-01-06 #2906 @huddlej
Thank you for the citation patch, @majensen! I've applied that and generated [a DOI for Augur from Zenodo](https://doi.org/10.5281/zenodo.4421955):

10.5281/zenodo.4421955
