### 2021-01-07 #2906 @arfon
@dcnickle, @Maghnuso - many thanks for your reviews here and to @majensen for editing this submission. JOSS relies upon the volunteer efforts of folks like yourselves, and we simply couldn't do this without you! ✨
