### 2021-01-07 #2906 @majensen
@huddlej thanks very much for choosing JOSS for this very nice work. Keep on with the outstanding efforts you and your colleagues are making; it is making a difference in the real world.
